Revert "t6006: drop "always" color config tests"
This reverts commit d66b4329220e81c64eb063adafa4e01e760afbce.

That commit was done primarily to prepare for the weakening
of "always" in 64c71c4615 (color: make "always" the same as
"auto" in config, 2017-10-03). But since we've now reverted
64c71c4615, there's no need for us to remove "-c
color.ui=always" from the tests. And in fact it's a good
idea to restore these tests, to make sure that "always"
continues to work.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>